Implement chat message history

+++ src/main/java/com/goolbitg/api/v1/repository/ChatMessageRepository.java
@@ -2,9 +2,9 @@ package com.goolbitg.api.v1.repository;
 
 import java.util.List;
 
-import org.apache.ibatis.annotations.Param;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 
 import com.goolbitg.api.v1.entity.chat.ChatMessage;
 
@@ -18,7 +18,7 @@ public interface ChatMessageRepository extends JpaRepository<ChatMessage, Long> 
     @Query("SELECT m FROM ChatMessage m WHERE m.buyOrNotId = :buyOrNotId AND m.id > :lastId ORDER BY m.id ASC")
     List<ChatMessage> findHistory(
         @Param("buyOrNotId") Long buyOrNotId,
-        @Param("lastId") int lastId
+        @Param("lastId") Long lastId
     );
 
 }
